Refactor Webpack config and dependencies to split vendor chunks
- Updated webpack.config.js to use granular splitChunks, isolating 'three' and 'interact' into their own chunks.
- Refactored src/utils/utils.ts to src/utils/three-utils.ts to decouple pure math from Three.js dependencies.
- Updated all HTML files in dist/ to include new script tags for three.js and interact.js.
- Verified that entry points like mathaven no longer bundle Three.js.
- All tests pass.
