+++ src/controller/rules/snooker.ts
@@ -13,7 +13,8 @@ import { Table } from "../../model/table"
 import { TableGeometry } from "../../view/tablegeometry"
 import { PlaceBall } from "../placeball"
 import { PlaceBallEvent } from "../../events/placeballevent"
-import { zero, isFirstShot } from "../../utils/utils"
+import { isFirstShot } from "../../utils/utils"
+import { zero } from "../../utils/three-utils"
 import { SnookerUtils, ShotInfo } from "./snookerutils"
 import { SnookerScoring } from "./snookerscoring"
 import { StartAimEvent } from "../../events/startaimevent"
