+++ src/controller/rules/threecushion.ts
@@ -12,7 +12,8 @@ import { Rack } from "../../utils/rack"
 import { CameraTop } from "../../view/cameratop"
 import { TableGeometry } from "../../view/tablegeometry"
 import { Rules } from "./rules"
-import { zero, isFirstShot } from "../../utils/utils"
+import { isFirstShot } from "../../utils/utils"
+import { zero } from "../../utils/three-utils"
 import { Respot } from "../../utils/respot"
 import { ThreeCushionConfig } from "../../utils/threecushionconfig"
 import { StartAimEvent } from "../../events/startaimevent"
