+++ src/events/placeballevent.ts
@@ -1,7 +1,7 @@
 import { GameEvent } from "./gameevent"
 import { EventType } from "./eventtype"
 import { Controller } from "../controller/controller"
-import { vec } from "../utils/utils"
+import { vec } from "../utils/three-utils"
 import { Vector3 } from "three"
 
 export interface RespotBody {
