+++ src/model/table.ts
@@ -9,7 +9,7 @@ import { TableGeometry } from "../view/tablegeometry"
 import { Outcome } from "./outcome"
 import { PocketGeometry } from "../view/pocketgeometry"
 import { bounceHanBlend } from "./physics/physics"
-import { zero } from "../utils/utils"
+import { zero } from "../utils/three-utils"
 import { R } from "./physics/constants"
 
 interface Pair {
