+++ src/utils/utils.ts
@@ -1,34 +1,8 @@
-import { Vector3 } from "three"
 import { EventType } from "../events/eventtype"
 import { Recorder } from "../events/recorder"
 
-export const zero = new Vector3(0, 0, 0)
-export const up = new Vector3(0, 0, 1)
-
 export function isFirstShot(recorder: Recorder): boolean {
   return !recorder.entries.some((e) => e.event.type === EventType.AIM)
-}
-
-export function vec(v) {
-  return new Vector3(v.x, v.y, v.z)
-}
-
-const upCrossVec = new Vector3()
-export function upCross(v) {
-  return upCrossVec.copy(up).cross(v)
-}
-const normVec = new Vector3()
-export function norm(v) {
-  return normVec.copy(v).normalize()
-}
-
-const vc = new Vector3()
-export function passesThroughZero(v, dv) {
-  return vc.copy(v).add(dv).dot(v) <= 0
-}
-
-export function unitAtAngle(theta, target = new Vector3()) {
-  return target.set(1, 0, 0).applyAxisAngle(up, theta)
 }
 
 export function round(num) {
@@ -38,20 +12,6 @@ export function round(num) {
 
 export function round2(num) {
   return Math.round((num + Number.EPSILON) * 100) / 100
-}
-
-export function roundVec(v) {
-  v.x = round(v.x)
-  v.y = round(v.y)
-  v.z = round(v.z)
-  return v
-}
-
-export function roundVec2(v) {
-  v.x = round2(v.x)
-  v.y = round2(v.y)
-  v.z = round2(v.z)
-  return v
 }
 
 export function atan2(y, x) {
