+++ src/view/aiminputs.ts
@@ -3,7 +3,7 @@ import { Container } from "../container/container"
 import { Input } from "../events/input"
 import { Session } from "../network/client/session"
 import { Overlap } from "../utils/overlap"
-import { unitAtAngle } from "../utils/utils"
+import { unitAtAngle } from "../utils/three-utils"
 import { id } from "../utils/dom"
 
 export class AimInputs {
