+++ test/events/eventutil.spec.ts
@@ -8,7 +8,7 @@ import { Table } from "../../src/model/table"
 import { Rack } from "../../src/utils/rack"
 import { BreakEvent } from "../../src/events/breakevent"
 import { PlaceBallEvent } from "../../src/events/placeballevent"
-import { zero } from "../../src/utils/utils"
+import { zero } from "../../src/utils/three-utils"
 import { ChatEvent } from "../../src/events/chatevent"
 import { BeginEvent } from "../../src/events/beginevent"
 import { HitEvent } from "../../src/events/hitevent"
