+++ test/model/table.spec.ts
@@ -3,7 +3,7 @@ import { Ball, State } from "../../src/model/ball"
 import { TableGeometry } from "../../src/view/tablegeometry"
 import { Table } from "../../src/model/table"
 import { Vector3 } from "three"
-import { zero } from "../../src/utils/utils"
+import { zero } from "../../src/utils/three-utils"
 import { Rack } from "../../src/utils/rack"
 import { R } from "../../src/model/physics/constants"
 import { PocketGeometry } from "../../src/view/pocketgeometry"
