+++ webpack.config.js
@@ -53,10 +53,23 @@ module.exports = {
     innerGraph: true,
     splitChunks: {
       cacheGroups: {
+        three: {
+          test: /[\\/]node_modules[\\/]three[\\/]/,
+          name: "three",
+          chunks: "all",
+          priority: 20,
+        },
+        interact: {
+          test: /[\\/]node_modules[\\/]interactjs[\\/]/,
+          name: "interact",
+          chunks: "all",
+          priority: 10,
+        },
         vendor: {
           test: /[\\/]node_modules[\\/]/,
           name: "vendor",
           chunks: "all",
+          priority: 0,
         },
       },
     },
